requirements.txt
torch==2.4.1
torchaudio==2.4.1
faster-whisper>=1.1.0
pyannote.audio>=3.3.0,<4.0.0
numpy>=1.24.0
av>=12.0.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
boto3>=1.34.0
